+++ extensions/quick-notes/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - New note metadata menu to show tags and other details for a note using `⌘ + M`
 - Sort straight from the notes list instead of preferences using `⌘ + S`'
-- Both these new features save your settings in LocalStorage
+- Both these new features save your settings in Cache
 
 ## [Copy Action] - 2024-05-08